@@ -1,36 +1,30 @@
 <?php
-/**
- * Phalcon Framework
- *
- * This source file is subject to the New BSD License that is bundled
- * with this package in the file docs/LICENSE.txt.
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@phalconphp.com so we can send you a copy immediately.
- *
- * PHP version 5
- *
- * @category Phalcon
- * @package  Phalcon\Annotations\Adapter
- * @author   Ilya Gusev <mail@igusev.ru>
- * @license  New BSD License
- * @link     http://phalconphp.com/
- */
+/*
+  +------------------------------------------------------------------------+
+  | Phalcon Framework                                                      |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2011-2015 Phalcon Team (http://www.phalconphp.com)       |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file docs/LICENSE.txt.                        |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+  | Authors: Ilya Gusev <mail@igusev.ru>                                   |
+  +------------------------------------------------------------------------+
+*/
 
 namespace Phalcon\Annotations\Adapter;
 
 use Phalcon\Annotations\Adapter;
-use Phalcon\Annotations\AdapterInterface;
 
 /**
  * \Phalcon\Annotations\Adapter\Base
  * Base class for \Phalcon\Annotations\Adapter\Memcached and other adapters.
  *
- * @category Phalcon
- * @package  Phalcon\Annotations\Adapter
- * @author   Ilya Gusev <mail@igusev.ru>
- * @license  New BSD License
- * @link     http://phalconphp.com/
+ * @package Phalcon\Annotations\Adapter
  */
 abstract class Base extends Adapter
 {
