@@ -3,7 +3,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2012 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2015 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -17,6 +17,7 @@
   |          Nikita Vershinin <endeveit@gmail.com>                         |
   +------------------------------------------------------------------------+
 */
+
 namespace Phalcon\Cache\Backend;
 
 use Phalcon\Cache\Exception;
@@ -30,7 +31,6 @@ use Phalcon\Db\AdapterInterface as DbAdapterInterface;
  */
 class Database extends Prefixable
 {
-
     /**
      * @var \Phalcon\Db\AdapterInterface
      */
