@@ -3,7 +3,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Framework                                                      |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2012 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2015 Phalcon Team (http://www.phalconphp.com)       |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file docs/LICENSE.txt.                        |
@@ -15,6 +15,7 @@
   | Authors: Nikita Vershinin <endeveit@gmail.com>                         |
   +------------------------------------------------------------------------+
 */
+
 namespace Phalcon\Cache\Backend;
 
 use Phalcon\Cache\Backend;
@@ -26,7 +27,6 @@ use Phalcon\Cache\BackendInterface;
  */
 abstract class Prefixable extends Backend implements BackendInterface
 {
-
     /**
      * Returns prefixed identifier.
      *
