@@ -13,7 +13,7 @@ use Phalcon\Loader\PSR as PsrLoader;
  *
  * @copyright (c) 2011-2015 Phalcon Team
  * @link      http://www.phalconphp.com
- * @author    Serghei Iakovlev <sadhooklay@gmail.com>
+ * @author    Serghei Iakovlev <serghei@phalconphp.com>
  * @package   Phalcon\Loader
  *
  * The contents of this file are subject to the New BSD License that is
